+++ uax29/RuneTokenizer.cs
@@ -19,25 +19,6 @@ internal static class RuneTokenizer
 	}
 }
 
-internal static class Decoders
-{
-	internal readonly static Decoders<char> Char = new(Rune.DecodeFromUtf16, Rune.DecodeLastFromUtf16);
-	internal readonly static Decoders<byte> Utf8 = new(Rune.DecodeFromUtf8, Rune.DecodeLastFromUtf8);
-}
-
-internal class Decoders<T>
-{
-	internal Decoder<T> FirstRune;
-	internal Decoder<T> LastRune;
-
-	internal Decoders(Decoder<T> firstRune, Decoder<T> lastRune)
-	{
-		this.FirstRune = firstRune;
-		this.LastRune = lastRune;
-	}
-}
-
-
 /// <summary>
 /// Tokenizer splits strings or UTF-8 bytes as words, sentences or graphemes, per the Unicode UAX #29 spec.
 /// </summary>
